refactor(tts): migrate OpenAI TTS to shared TextToSpeechModel interface and standardize on Double for speed
  BREAKING CHANGE: OpenAI TTS now uses shared interfaces instead of provider-specific classes

  - Replace deprecated OpenAI-specific classes (SpeechModel, SpeechPrompt, SpeechResponse,
    StreamingSpeechModel) with shared interfaces (TextToSpeechModel, TextToSpeechPrompt,
    TextToSpeechResponse, StreamingTextToSpeechModel)
  - Update interface hierarchy: TextToSpeechModel now extends StreamingTextToSpeechModel
    for consistency with ChatModel pattern
  - Add @FunctionalInterface annotation to StreamingTextToSpeechModel
  - Migrate speed property from Float to Double across all OpenAI TTS components for
    consistency with ElevenLabs API and better type uniformity
  - Update OpenAiAudioSpeechOptions to implement TextToSpeechOptions interface
  - Add TextToSpeechModelTests with 7 comprehensive unit tests for interface behavior
  - Improve null safety in TextToSpeechModel.call(String) default method
  - Update documentation to reflect new interface hierarchy and Double usage
  - Add migration guide to documentation for upgrading from deprecated classes

+++ auto-configurations/models/spring-ai-autoconfigure-model-openai/src/main/java/org/springframework/ai/model/openai/autoconfigure/OpenAiAudioSpeechProperties.java
@@ -38,7 +38,7 @@ public class OpenAiAudioSpeechProperties extends OpenAiParentProperties {
 
 	public static final String DEFAULT_SPEECH_MODEL = OpenAiAudioApi.TtsModel.GPT_4_O_MINI_TTS.getValue();
 
-	private static final Float SPEED = 1.0f;
+	private static final Double SPEED = 1.0;
 
 	private static final String VOICE = OpenAiAudioApi.SpeechRequest.Voice.ALLOY.getValue();
